Reorganizes the CDN tests.
