+++ tests/Unit/inc/classes/CDN/CDN/TestCase.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace WP_Rocket\Tests\Unit\inc\classes\CDN\CDN;
+
+use Brain\Monkey\Functions;
+use WPMedia\PHPUnit\Unit\TestCase as BaseTestCase;
+
+abstract class TestCase extends BaseTestCase {
+
+	protected function setUp() {
+		parent::setUp();
+
+		Functions\when( 'wp_parse_url' )->alias( function( $url, $component = -1 ) {
+			return parse_url( $url, $component );
+		} );
+	}
+
+	protected function getOptionsMock( array $map = [] ) {
+		$options = $this->createMock( 'WP_Rocket\Admin\Options_Data' );
+
+		if ( empty( $map ) ) {
+			$map = [
+				[ 'cdn', '', 1 ],
+				[ 'cdn_cnames', [], [ 'cdn.example.org' ] ],
+				[ 'cdn_reject_files', [], [] ],
+				[ 'cdn_zone', [], [ 'all' ] ],
+			];
+		}
+		$options->method( 'get' )->will( $this->returnValueMap( $map ) );
+
+		return $options;
+	}
+}
