+++ tests/Unit/inc/classes/CDN/RocketCDN/CDNOptionsManager/disable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\CDN\RocketCDN\CDNOptionsManager;
+namespace WP_Rocket\Tests\Unit\inc\classes\CDN\RocketCDN\CDNOptionsManager;
 
 use Brain\Monkey\Functions;
 use WPMedia\PHPUnit\Unit\TestCase;
@@ -9,21 +9,20 @@ use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\CDN\RocketCDN\CDNOptionsManager;
 
 /**
- * @covers\WP_Rocket\CDN\RocketCDN\CDNOptionsManager::enable
+ * @covers\WP_Rocket\CDN\RocketCDN\CDNOptionsManager::disable
  * @group RocketCDN
  */
-class Test_Enable extends TestCase {
-	public function testShouldEnableCDNOptions() {
+class Test_Disable extends TestCase {
+	public function testShouldDisableCDNOptions() {
 		$expected = [
-			'cdn'        => 1,
-			'cdn_cnames' => [
-				'https://rocketcdn.me',
-			],
-			'cdn_zone'   => [
-				'all',
-			],
+			'cdn'        => 0,
+			'cdn_cnames' => [],
+			'cdn_zone'   => [],
 		];
 
+		Functions\expect( 'delete_option' )
+			->once()
+			->with( 'rocketcdn_user_token' );
 		Functions\expect( 'delete_transient' )
 			->once()
 			->with( 'rocketcdn_status' );
@@ -32,9 +31,9 @@ class Test_Enable extends TestCase {
 		$options_array->expects( $this->exactly( 3 ) )
 		              ->method( 'set' )
 		              ->withConsecutive(
-			              [ 'cdn', 1 ],
-			              [ 'cdn_cnames', [ 'https://rocketcdn.me' ] ],
-			              [ 'cdn_zone', [ 'all' ] ]
+			              [ 'cdn', 0 ],
+			              [ 'cdn_cnames', [] ],
+			              [ 'cdn_zone', [] ]
 		              );
 		$options_array->method( 'get_options' )
 		              ->willReturn( $expected );
@@ -47,6 +46,6 @@ class Test_Enable extends TestCase {
 		( new CDNOptionsManager(
 			$options,
 			$options_array
-		) )->enable( 'https://rocketcdn.me' );
+		) )->disable();
 	}
 }
