+++ tests/Unit/inc/classes/CDN/RocketCDN/CDNOptionsManager/enable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\CDN\RocketCDN\CDNOptionsManager;
+namespace WP_Rocket\Tests\Unit\inc\classes\CDN\RocketCDN\CDNOptionsManager;
 
 use Brain\Monkey\Functions;
 use WPMedia\PHPUnit\Unit\TestCase;
@@ -9,20 +9,21 @@ use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\CDN\RocketCDN\CDNOptionsManager;
 
 /**
- * @covers\WP_Rocket\CDN\RocketCDN\CDNOptionsManager::disable
+ * @covers\WP_Rocket\CDN\RocketCDN\CDNOptionsManager::enable
  * @group RocketCDN
  */
-class Test_Disable extends TestCase {
-	public function testShouldDisableCDNOptions() {
+class Test_Enable extends TestCase {
+	public function testShouldEnableCDNOptions() {
 		$expected = [
-			'cdn'        => 0,
-			'cdn_cnames' => [],
-			'cdn_zone'   => [],
+			'cdn'        => 1,
+			'cdn_cnames' => [
+				'https://rocketcdn.me',
+			],
+			'cdn_zone'   => [
+				'all',
+			],
 		];
 
-		Functions\expect( 'delete_option' )
-			->once()
-			->with( 'rocketcdn_user_token' );
 		Functions\expect( 'delete_transient' )
 			->once()
 			->with( 'rocketcdn_status' );
@@ -31,9 +32,9 @@ class Test_Disable extends TestCase {
 		$options_array->expects( $this->exactly( 3 ) )
 		              ->method( 'set' )
 		              ->withConsecutive(
-			              [ 'cdn', 0 ],
-			              [ 'cdn_cnames', [] ],
-			              [ 'cdn_zone', [] ]
+			              [ 'cdn', 1 ],
+			              [ 'cdn_cnames', [ 'https://rocketcdn.me' ] ],
+			              [ 'cdn_zone', [ 'all' ] ]
 		              );
 		$options_array->method( 'get_options' )
 		              ->willReturn( $expected );
@@ -46,6 +47,6 @@ class Test_Disable extends TestCase {
 		( new CDNOptionsManager(
 			$options,
 			$options_array
-		) )->disable();
+		) )->enable( 'https://rocketcdn.me' );
 	}
 }
